+++ rss-client/src/app/Quiz/employee/test-in-progress/test-in-progress.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { TestInProgressComponent } from './test-in-progress.component';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { QuizService } from 'src/app/services/quiz.service';
+import { QuizService } from 'src/app/Quiz/service/quiz.service';
 import { RouterTestingModule } from '@angular/router/testing';
 
 describe('TestInProgressComponent', () => {});
